allow custom format on item for Debug
